+++ saleor/graphql/channel/mutations/channel_update.py
@@ -110,6 +110,12 @@ class ChannelUpdate(ModelMutation):
                     "automatically_fulfill_non_shippable_gift_card"
                 ] = automatically_fulfill_non_shippable_gift_card
 
+            if "expire_orders_after" in order_settings:
+                expire_orders_after = order_settings["expire_orders_after"]
+                cleaned_input["expire_orders_after"] = cls.clean_expire_orders_after(
+                    expire_orders_after
+                )
+
             if mark_as_paid_strategy := order_settings.get("mark_as_paid_strategy"):
                 cleaned_input["order_mark_as_paid_strategy"] = mark_as_paid_strategy
 
@@ -120,6 +126,21 @@ class ChannelUpdate(ModelMutation):
                     "default_transaction_flow_strategy"
                 ] = default_transaction_strategy
         return cleaned_input
+
+    @classmethod
+    def clean_expire_orders_after(cls, expire_orders_after):
+        if expire_orders_after is None or expire_orders_after == 0:
+            return None
+        if expire_orders_after < 0:
+            raise ValidationError(
+                {
+                    "expire_orders_after": ValidationError(
+                        "Expiration time for orders cannot be lower than 0.",
+                        code=ChannelErrorCode.INVALID.value,
+                    )
+                }
+            )
+        return expire_orders_after
 
     @classmethod
     def check_permissions(cls, context, permissions=None, **data):