+++ saleor/graphql/channel/tests/queries/test_channel.py
@@ -288,6 +288,7 @@ QUERY_CHANNEL_ORDER_SETTINGS = """
             orderSettings {
                 automaticallyConfirmAllNewOrders
                 automaticallyFulfillNonShippableGiftCard
+                expireOrdersAfter
                 markAsPaidStrategy
                 defaultTransactionFlowStrategy
             }
@@ -331,6 +332,10 @@ def test_query_channel_order_settings_as_staff_user(
         == channel_USD.automatically_fulfill_non_shippable_gift_card
     )
     assert (
+        channel_data["orderSettings"]["expireOrdersAfter"]
+        == channel_USD.expire_orders_after
+    )
+    assert (
         channel_data["orderSettings"]["defaultTransactionFlowStrategy"]
         == channel_USD.default_transaction_flow_strategy.upper()
     )
@@ -371,6 +376,10 @@ def test_query_channel_order_settings_as_app(
     assert (
         channel_data["orderSettings"]["automaticallyFulfillNonShippableGiftCard"]
         == channel_USD.automatically_fulfill_non_shippable_gift_card
+    )
+    assert (
+        channel_data["orderSettings"]["expireOrdersAfter"]
+        == channel_USD.expire_orders_after
     )
     assert (
         channel_data["orderSettings"]["defaultTransactionFlowStrategy"]