+++ saleor/graphql/channel/types.py
@@ -25,6 +25,7 @@ from ..core.descriptions import (
     PREVIEW_FEATURE,
 )
 from ..core.fields import PermissionsField
+from ..core.scalars import Minute
 from ..core.types import CountryDisplay, ModelObjectType, NonNullList
 from ..meta.types import ObjectWithMetadata
 from ..translations.resolvers import resolve_translation
@@ -180,6 +181,14 @@ class OrderSettings(ObjectType):
         description=(
             "When enabled, all non-shippable gift card orders "
             "will be fulfilled automatically."
+        ),
+    )
+    expire_orders_after = Minute(
+        required=False,
+        description=(
+            "Expiration time in minutes. Default null - means do not expire any orders."
+            + ADDED_IN_313
+            + PREVIEW_FEATURE
         ),
     )
 
@@ -451,6 +460,7 @@ class Channel(ModelObjectType):
             automatically_fulfill_non_shippable_gift_card=(
                 root.automatically_fulfill_non_shippable_gift_card
             ),
+            expire_orders_after=root.expire_orders_after,
             mark_as_paid_strategy=root.order_mark_as_paid_strategy,
             default_transaction_flow_strategy=root.default_transaction_flow_strategy,
         )