+++ saleor/graphql/payment/tests/mutations/test_transaction_create.py
@@ -7,7 +7,7 @@ from mock import patch
 from .....checkout import CheckoutAuthorizeStatus, CheckoutChargeStatus
 from .....checkout.calculations import fetch_checkout_data
 from .....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
-from .....order import OrderAuthorizeStatus, OrderEvents
+from .....order import OrderAuthorizeStatus, OrderEvents, OrderStatus
 from .....order.utils import update_order_authorize_data, update_order_charge_data
 from .....payment import TransactionEventStatus, TransactionEventType
 from .....payment.error_codes import TransactionCreateErrorCode
@@ -1592,6 +1592,58 @@ def test_creates_transaction_event_for_checkout_by_staff(
     assert event.app_identifier is None
     assert event.app is None
     assert event.type == TransactionEventType.INFO
+
+
+def test_creates_transaction_automatically_confirm(
+    order_with_lines, permission_manage_payments, app_api_client
+):
+    # given
+    order_with_lines.status = OrderStatus.UNCONFIRMED
+    order_with_lines.save()
+    order_with_lines.channel.automatically_confirm_all_new_orders = True
+    order_with_lines.channel.save()
+
+    status = "Failed authorized for 10$"
+    type = "Credit Card"
+    reference = "PSP reference - 123"
+    available_actions = []
+    authorized_value = Decimal("0")
+    metadata = {"key": "test-1", "value": "123"}
+    private_metadata = {"key": "test-2", "value": "321"}
+
+    event_reference = "PSP-ref"
+    event_name = "Test"
+    variables = {
+        "id": graphene.Node.to_global_id("Order", order_with_lines.pk),
+        "transaction": {
+            "status": status,
+            "type": type,
+            "reference": reference,
+            "availableActions": available_actions,
+            "amountAuthorized": {
+                "amount": authorized_value,
+                "currency": "USD",
+            },
+            "metadata": [metadata],
+            "privateMetadata": [private_metadata],
+        },
+        "transaction_event": {
+            "status": TransactionEventStatusEnum.SUCCESS.name,
+            "reference": event_reference,
+            "name": event_name,
+        },
+    }
+
+    # when
+    response = app_api_client.post_graphql(
+        MUTATION_TRANSACTION_CREATE, variables, permissions=[permission_manage_payments]
+    )
+
+    # then
+    get_graphql_content(response)
+
+    order_with_lines.refresh_from_db()
+    assert order_with_lines.status == OrderStatus.UNFULFILLED
 
 
 def test_transaction_create_external_url_incorrect_url_format_by_app(