+++ saleor/plugins/webhook/tests/subscription_webhooks/test_create_deliveries_for_subscription.py
@@ -1191,6 +1191,22 @@ def test_order_cancelled(order, subscription_order_cancelled_webhook):
     assert deliveries[0].webhook == webhooks[0]
 
 
+def test_order_expired(order, subscription_order_expired_webhook):
+    # given
+    webhooks = [subscription_order_expired_webhook]
+    event_type = WebhookEventAsyncType.ORDER_EXPIRED
+    order_id = graphene.Node.to_global_id("Order", order.id)
+    expected_payload = json.dumps({"order": {"id": order_id}})
+
+    # when
+    deliveries = create_deliveries_for_subscriptions(event_type, order, webhooks)
+
+    # then
+    assert deliveries[0].payload.payload == expected_payload
+    assert len(deliveries) == len(webhooks)
+    assert deliveries[0].webhook == webhooks[0]
+
+
 def test_order_fulfilled(order, subscription_order_fulfilled_webhook):
     webhooks = [subscription_order_fulfilled_webhook]
     event_type = WebhookEventAsyncType.ORDER_FULFILLED